# laravel-app/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/admin/engineers": {
            "get": {
                "tags": [
                    "Engineers"
                ],
                "summary": "Obtenir la liste des ingénieurs",
                "operationId": "getEngineers",
                "responses": {
                    "200": {
                        "description": "Liste des ingénieurs récupérée avec succès"
                    },
                    "500": {
                        "description": "Erreur interne"
                    }
                }
            },
            "post": {
                "tags": [
                    "Engineers"
                ],
                "summary": "Enregistrer un nouvel ingénieur",
                "operationId": "createEngineer",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Engineer's name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Engineer's email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Ingénieur créé avec succès"
                    },
                    "422": {
                        "description": "Données invalides"
                    }
                }
            }
        },
        "/api/admin/engineers/{engineer_id}": {
            "get": {
                "tags": [
                    "Engineers"
                ],
                "summary": "Afficher les détails d'un ingénieur spécifique",
                "operationId": "getEngineer",
                "parameters": [
                    {
                        "name": "engineer_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ingénieur récupéré avec succès"
                    },
                    "404": {
                        "description": "Ingénieur non trouvé"
                    }
                }
            },
            "put": {
                "tags": [
                    "Engineers"
                ],
                "summary": "Mettre à jour un ingénieur existant",
                "operationId": "updateEngineer",
                "parameters": [
                    {
                        "name": "engineer_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Nom de l'ingénieur",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "Email de l'ingénieur",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ingénieur mis à jour avec succès"
                    },
                    "422": {
                        "description": "Données invalides"
                    },
                    "404": {
                        "description": "Ingénieur non trouvé"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Engineers"
                ],
                "summary": "Supprimer un ingénieur existant",
                "operationId": "deleteEngineer",
                "parameters": [
                    {
                        "name": "engineer_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ingénieur supprimé avec succès"
                    },
                    "404": {
                        "description": "Ingénieur non trouvé"
                    }
                }
            }
        },
        "/api/admin/accept/plan/{plan_id}": {
            "post": {
                "tags": [
                    "Plans"
                ],
                "summary": "Accepter ou rejeter un plan créé par un ingénieur",
                "operationId": "acceptPlan",
                "parameters": [
                    {
                        "name": "plan_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "comment",
                                    "accept"
                                ],
                                "properties": {
                                    "comment": {
                                        "description": "Commentaire sur la décision",
                                        "type": "string"
                                    },
                                    "accept": {
                                        "description": "Accepter ou rejeter le plan",
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Décision prise avec succès"
                    },
                    "422": {
                        "description": "Données invalides"
                    },
                    "404": {
                        "description": "Plan non trouvé"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}